Support aliased Mongoid fields with options[:as] parameter. The options[:field] paramter is no longer externally accessible.
For example:

taggable :t, as: tags

Will store the tags array in a database field 't', but all it's methods will use the names tags, e.g. 'tags_string', etc. This works because :as is delegated to Mongoid, which uses this alias feature.

This is useful to reduce storage size in MongoDB.

This commit includes some other trivial changes (TAGGABLE_DEFAULT_SEPARATOR is now DEFAULT_SEPARATOR)